starting refactor
