+++ src/h3/api/basic_int/_api_template.py
