+++ src/h3/api/memview_int/_api_template.py
@@ -1,0 +1,1 @@
+../basic_int/_api_template.py
