--- src/h3/api/memview_int/_public_api.py
@@ -1,1 +1,0 @@
-../basic_int/_public_api.py
